Refactor: AdminTool component refactored
